kvizListAdapter - prikaz kvizova OK + moguce uci i raditi ih

+++ app/src/main/java/ba/etf/rma21/projekat/data/models/OdgovorBody.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 class OdgovorBody(
         @SerializedName("odgovor") val odgovor: Int,
         @SerializedName("pitanje") val pitanje: Int,
-        @SerializedName("bodovi") val bodovi: Int
+        @SerializedName("bodovi") val bodovi: Float
 
 
 )
